+++ src/irc/outputfilter.h
@@ -62,15 +62,15 @@ namespace Konversation
             OutputFilterResult parse(const QString& myNick,const QString& line,const QString& name);
 
             // dcc send
-            OutputFilterResult sendRequest(const QString &recipient,const QString &fileName,const QString &address,uint port,unsigned long size);
-            OutputFilterResult passiveSendRequest(const QString& recipient,const QString &fileName,const QString &address,unsigned long size,const QString &token);
-            OutputFilterResult acceptResumeRequest(const QString &recipient,const QString &fileName,uint port,int startAt);
-            OutputFilterResult acceptPassiveResumeRequest(const QString &recipient,const QString &fileName,uint port,int startAt,const QString &token);
+            OutputFilterResult sendRequest(const QString &recipient,const QString &fileName,const QString &address,uint port,quint64 size);
+            OutputFilterResult passiveSendRequest(const QString& recipient,const QString &fileName,const QString &address,quint64 size,const QString &token);
+            OutputFilterResult acceptResumeRequest(const QString &recipient,const QString &fileName,uint port,quint64 startAt);
+            OutputFilterResult acceptPassiveResumeRequest(const QString &recipient,const QString &fileName,uint port,quint64 startAt,const QString &token);
 
             // dcc recv
             OutputFilterResult resumeRequest(const QString &sender,const QString &fileName,uint port,KIO::filesize_t startAt);
             OutputFilterResult resumePassiveRequest(const QString &sender,const QString &fileName,uint port,KIO::filesize_t startAt,const QString &token);
-            OutputFilterResult acceptPassiveSendRequest(const QString& recipient,const QString &fileName,const QString &address,uint port,unsigned long size,const QString &token);
+            OutputFilterResult acceptPassiveSendRequest(const QString& recipient,const QString &fileName,const QString &address,uint port,quint64 size,const QString &token);
             OutputFilterResult rejectDccSend(const QString& partnerNick, const QString& fileName);
 
             // dcc chat
